requirements.txt
streamlit
openai
anthropic
google-generativeai
pydantic
python-dotenv
tenacity
tldextract
requests
urllib3
beautifulsoup4
lxml_html_clean
newspaper3k
selenium
webdriver-manager
pypdf
python-docx
filetype
charset-normalizer
pandas
plotly
tomli
